scripts: Add some scripts to dump spreadsheets for Endless Content
dump_apps.py creates a CSV file of:

    itemId, appId, titleWord1, titleWord2, titleWord3, tag1, tag2, tag3

for each record of each shard of each app, which can be used later
do to some learning on.

create_access.py creates a CSV file of itemId, timestamp

where timestamp is the unix time of last access for itemId . This can
be used later to learn which articles are relevant to the users
interests at a given point in time.